+++ frontend/src/components/customer_analytics/pages/index.ts
@@ -3,3 +3,4 @@ export { default as CustomerOverview } from './CustomerOverview';
 export { default as CustomerDemographics } from './CustomerDemographics';
 export { default as CustomerBehavior } from './CustomerBehavior';
 export { default as CustomerSegmentation } from './CustomerSegmentation';
+export { default as CustomerValueAnalytics } from './CustomerValueAnalytics';